refactor codegen scripts to make it easier to generate two clients
