+++ pkg/client/clientset/versioned/fake/register.go
@@ -34,10 +34,10 @@ var scheme = runtime.NewScheme()
 var codecs = serializer.NewCodecFactory(scheme)
 
 var localSchemeBuilder = runtime.SchemeBuilder{
+	gatewayv1.AddToScheme,
 	gatewayv1alpha2.AddToScheme,
 	gatewayv1alpha3.AddToScheme,
 	gatewayv1beta1.AddToScheme,
-	gatewayv1.AddToScheme,
 }
 
 // AddToScheme adds all types of this clientset into the given scheme. This allows composition
